requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pillow==10.3.0
opencv-python==4.9.0.80
imagehash==4.3.1
numpy==1.26.4
python-multipart==0.0.9
jinja2==3.1.4
pydantic==2.6.4
